WIP external auth server support
This is a proof-of-concept hack. The goal is to wire up enough machinery
that we can make one or two authentication servers work. This process
should help us figure out what the final API needs to look like.

Add a SupportService CRD and plumb it through Contour so that it generates
a cluster that can be used to reach a GRPC service.

Add some extra fields to the HTTPProxy virtual host that specify a
reference to a SupportService to wire up as an external auth server.
This is where `ext_authz` parameters will end up going.

Client auth requires a TLS virtual host, and is not compatible with
certificate fallbacks. The reason for this is that the authz server is
a property of the HTTP connection manager, and only TLS virtual hosts
get their own HCM.

A working example:

apiVersion: cert-manager.io/v1alpha2
kind: Certificate
metadata:
  name: echo
spec:
  dnsNames:
  - echo.projectcontour.io
  secretName: echo
  issuerRef:
    name: selfsigned

---

apiVersion: projectcontour.io/v1alpha1
kind: SupportService
metadata:
  name: echo
spec:
  service:
    name: echo
    port: 80

---

apiVersion: projectcontour.io/v1
kind: HTTPProxy
metadata:
  name: echo
spec:
  virtualhost:
    fqdn: echo.projectcontour.io
    tls:
      secretName: echo
    clientAuthenticationPolicy:
      supportServiceRef:
        apiVersion: projectcontour.io/v1alpha1
        kind: SupportService
        name: echo
  routes:
  - services:
    - name: echo
      port: 80