internal/k8s: consistently use the controller-runtime informer cache
Prior to this change, the Kubernetes client API exposed methods to
create a informer factories as well as namespaced informer factories.
This encouraged other systems to use multiple factories and multiple
informers for each factory. The result of this is that objects could be
cached multiple times in the different informer factory caches, and that
the informer states could be subtly out of sync.

This change removes the internal APIs that expose informer factories.
There is now just one API to create an informer, and it is backed by the
controller-runtime cache API. This ensures that all the informers used
across Contour share a common cache and therefore a consistent view of
Kubernetes state.

This updates #2683.
This fixes #2904.

Signed-off-by: James Peach <jpeach@vmware.com>